Task #4: Checkpoint B 리뷰 수정 3건 반영
- xref.rs: parse_xref_chain에서 visited 검사를 depth 검사보다 먼저 수행
  (100개 순환 chain에서 XrefChainCycle 대신 XrefChainTooDeep이 반환되는 명세 위반 수정)
- xref.rs: consume_newline에서 \r 단독 거부 (엄격 EOL 정책)
- xref_tests.rs: inline #[cfg(test)]를 tests/parser/xref_tests.rs로 이동,
  회귀 테스트 2건 추가 (100개 순환 chain / 101개 비순환 chain)
- xref-chain-check-order.md: 트러블슈팅 문서 신규 작성

closes #4 (Checkpoint B 리뷰)
